Refactor and add support for config file
Create GeneratorEnvrionment class to hold "global" variables such as paths, config, jinja environment

Add support for config file in TOML format. Can be specified with --config, and if not specified we'll look for it in ./config/rayter.toml and ./rayter.toml.

Add support for specifying website title in config file.